Increase the padding from both ends on the top nav bar. Also I want the logo to be exactly like how I had uploaded.
I want you to comment out the codes for section 2, 3, 4, 5, 6 and only have the hero section, the contact us section and the footer. In the footer please change the Book a demo section to join waitlist section.

In the footer section please comment out the social media icons, also remove the entire sections except the terms of service and the privacy policy.