Change the logo to the one I just uploaded. Change the Book a demo button to Join waitlist in hero. Also remove both the buttons in the top nav bar.
In the contact us section:
change the email to contact@lexamplify.com
Location to New Delhi.
Change the background of the contact details card to be that of the primary color.
remove the phone number. Make the length of the left card to be the same as that of the right card.